[FreeRTOS] enable MMU for FreeRTOS QEMU
This commit sets the translation table, for all the address spaces (
be it, the DRAM region allocated for freeRTOS, Device I/O region). It
demonstrates two approaches to create the address translation for
needed address space.

First approach, make use of the simplest ever set-up with only a
level 1 translation table (of 3 levels of tables at level 1, 2 and 3).
Each entry in this table (level 1) covers 1GB of physical address space,
and we are using entry 0 and entry 2 to cover the needed address space.

Second approach, make use of 2 level of translation (at level1 and level2),
where table walk starts from level 1 table, and two tables at level2 maps
the needed regions. Advantage of using, this approach (and is enabled here)
over the first one is that memory attributes are set for pages on a small
granularity level, .i.e. 2MB.

Signed-off-by: Amit Singh Tomar <atomar25opensource@gmail.com>